Implemented SciJavaPlugin format of JEX functions
Created the annotations for inputs, parameters, and outputs of
JEXPlugin objects.
Used the ImageJ2 PluginService to auto discover the plugins.
Created a JEXCrunchablePlugin which wraps a JEXPlugin with an
JEXCrunchable interface so it automatically plugs into the normal batch
paradigm.
Renamed packages in light of changes to keep things better organized as
we fade out old plugins and replace them with new ones.

TODO: Make writers that don't take a name as an input parameter to the
method.