Fix bug in starting up DefaultLegacyService
Before, if you opened an image prior to opening the processing panel,
the old IJ stuff would be loaded prior to IJ2 stuff so you get an
error. Calling the static DefaultLegacyService.preinit() gets rid of
this error but it must be called before any opportunity for calling old
IJ stuff. So, it must be moved out of IJPluginUtility to the main class
of JEX.