Add a toplevel application module
This module is intended to be a single dependency for anyone who wants
to depend on JEX, as well as a main method entry point.

It declares the core JEX library itself (jex-core) as a compile scope
dependency, and declares all the optional plugins as runtime scope so
that they are available (even though they are technically not needed to
actually build the code).