Explicitly override scijava-common version
For some reason, Maven was resolving scijava-common's version as 2.24.0
instead of 2.27.0, despite the fact that 2.27.0 is the version declared
in the relevant version of pom-scijava. It seems that 2.24.0 is
referenced transitively from certain other dependency versions, but it
is odd that the conflict resolution is preferring it.

So this change fixes things "for now" but of course when the version of
pom-imagej is bumped, the scijava-common.version property override
should be attempted to be removed, to see if the problem goes away.
Otherwise, JEX will be "stuck" at 2.27.0 even when pom-imagej etc. has
moved on, which might ultimately cause more JAR version skew
(NoSuchMethodError etc.).