Fix uncaught failures in DB saving and creation
Catch previously uncaught failures to created necessary folders. Also
when creating a database, first test if the repository into which you
would like to created the database is connected to this computer. Also,
now check to see if the DBInfo file is saved when creating a database.
Prior to this, you could fail to create the DBInfo and not know it
because none of the "string" and "xml" and "DBInfo" saving methods
relayed success of saving or not.