Add ability to import ND2 Datasets directly
Needed to import scifio and scifio-bf-compat
Tested it in PointTester
Created a ui dialog on the Script panel of JEX
Created and associated JEXPlugin that takes file paths in a Value
Object to feed in multiple file paths to import many ND2 files at once.

Updated capitalization of Static string variable names in
MarkerConstants to be more consistent.