Partially fix ND2 file capabilities
The way to depend on scifio changed. We must now specify the gpl and
bsd jars explicitly as a dependency. They used to be brought in
automatically. This is a partial fix because we still need to upgrade
to the next formats versions 5.1.x to fix some metadata capabilities
for ND2 formats. That will be maintained on a separate branch until
release of 5.1.x and the corresponding incorporation of it into
scifio-bf-compat 1.11.x (which is currently just a branch itself).