Make name transfer of colors optional
Ran into an issue with an example ND2 file where the number of color
names was different than the number of colors taken in the image. This
screwed up indexing. In these situations, not sure what is best to do
so I leave the option to just leave the indices.

Also didn't realize the key set of the MetaTable wasn't ordered, so it
is now ordered by Name # to hopefully match up with the order of the
colors as they are provided by the plane interator.