Created PropertySheetRB.properties resource
For some reason, the preferences pane stopped working unless this was
included as a resource, otherwise you got a resource not found error.
Eclipse was keeping it in the src/main/java arm of the project when it
needed to be in the src/main/resources arm for maven. Don't know why we
didn't run into it earlier.