Enable outputs of JEXData collections and maps
This was needed for the ability to split an image based on a dimension.
Given the number of outputs is variable, we needed a way to have a
variable number of outputs but the output shouldn't show up in the gui
given the output names are determined by the function implementation.
So, we...

1) only add outputs that extend JEXData
2) collect any JEXData that are contained within collections or maps
that are marked with the @OutputParameterMarker annotation.