Avoid static initialization of ImageJ instance
This caused issues. If legacy ij was used before new imageJ, then we
couldn't discover plugins and we couldn't initialize CrunchFactory.
Seems to be related to the static initialization of the ImageJ instance
during the initialization of IJ2PluginUtility and CrunchFactory which
needed the IJ2PluginUtility. By creating getters for the ij instance
and for the TreeMap of JEXCrunchables, we could more safely initialize
the classes and catch errors, which can't be done in static
initialization blocks etc (which was the unsafe part of how we were
doing things as we couldn't catch and recover, causing a domino
effect).