Finished adding and testing functions for adding and removing singleton dimensions
The functions allow the user to add a single dimension that will
contain a single value or to remove a single dimension that contains a
single value. The functions do not show and output link on purpose as
this function must work with any type of object so we can't guarantee
the type of the output. The functions are located in the "JEX/Data
Editing" toolbox of functions in the processing pane.