Create intersection of label regions with masks
This allows us to define data based on whatever is inside the boundary
of a whole cell. This provides traceability of subregions to each cell.
However, right now there can only be one subregion per channel per
cell. For example, if staining for nuclei and two nuclei exist within a
single whole cell, the two nuclei will be labeled with the same number,
resulting in a composite subregion within the cell that might be weird.