Remove reference to Roi
This would result in a memory leak when you have a TreeMap<DimensionMap,
ROIPlus>. When you call RoiPlus.getRoi(), it would create the internal
reference. This Roi reference holds a reference to the
ImagePlus/ImageProcessor that it is applied to. So all the ROIPlus's
would then be keeping the images from being garbage collected.