Fix saving issue.
Files sometimes got stranded in the temp folder because the last file
tested for the JEXData was seen as being in the correct place. Now, any
file that is not in the right place is copied to the temp folder. This
affected things when renaming some but not all dimension values.